+++ models/authenticator/store/db_store.ts
@@ -2,57 +2,86 @@ import * as mysql from 'mysql';
 import { SDKException } from '../../../core/com/zoho/crm/api/exception/sdk_exception';
 import { UserSignature } from "../../../routes/user_signature";
 import { Constants } from "../../../utils/util/constants";
-import { OAuthToken, TokenType } from '../oauth_token';
+import { OAuthToken } from '../oauth_token';
 import { Token } from "../token";
 import { TokenStore } from "./token_store";
+import { OAuthBuilder } from "../oauth_builder";
 
 /**
  * This class stores the user token details to the MySQL DataBase.
 */
 export class DBStore implements TokenStore {
     private userName: string;
+
     private portNumber: number;
+
     private password: string;
+
     private host: string;
+
     private databaseName: string;
+
+    private tableName: string;
 
     /**
      * Creates an DBStore class instance with the specified parameters.
      * @param {string} host - A String containing the DataBase host name. Default value is localhost
      * @param {string} databaseName - A String containing the DataBase name. Default value is zohooauth
+     * @param {string} tableName - A String containing the DataBase table name. Default value is oauthtoken
      * @param {string} userName - A String containing the DataBase user name. Default value is root
      * @param {string} password - A String containing the DataBase password. Default value is an empty string
      * @param {string} portNumber - A String containing the DataBase port number. Default value is 3306
-    */
-    constructor(host?: string, databaseName?: string, userName?: string, password?: string, portNumber?: number) {
-        this.host = (host != undefined) ? host : Constants.MYSQL_HOST;
-        this.databaseName = (databaseName != undefined) ? databaseName : Constants.MYSQL_DATABASE_NAME;
-        this.userName = (userName != undefined) ? userName : Constants.MYSQL_USER_NAME;
-        this.password = (password != undefined) ? password : "";
-        this.portNumber = (portNumber != undefined) ? portNumber : Constants.MYSQL_PORT_NUMBER;
+     */
+    constructor(host: string, databaseName: string, tableName: string, userName: string, password: string, portNumber: number) {
+        this.host = host;
+
+        this.databaseName = databaseName;
+
+        this.tableName = tableName;
+
+        this.userName = userName;
+
+        this.password = password;
+
+        this.portNumber = portNumber;
     }
 
     async getToken(user: UserSignature, token: Token): Promise<Token | undefined> {
         try {
-            var connection = this.getConnection();
-            if(token instanceof OAuthToken) {
-                var sql = this.constructDBQuery(user.getEmail(), token, false);
+            var connection = await this.getConnection();
+
+            if (token instanceof OAuthToken) {
                 var oauthToken = token;
-                return new Promise(function(resolve, reject){
-                    connection.connect(function(err){
-                        if(err) throw err;
-                        connection.query(sql, function(err, result){
-                            if(err) {
+
+                var sql = await this.constructDBQuery(user.getEmail(), oauthToken, false);
+
+                return new Promise(function (resolve, reject) {
+                    connection.connect(function (err) {
+                        if (err) throw err;
+
+                        connection.query(sql, function (err, result) {
+                            if (err) {
                                 connection.end();
+
                                 throw err;
                             }
+
                             connection.end();
-                            if (result.length!=0){
+
+                            if (result.length != 0) {
+                                oauthToken.setId(result[0].id);
+
                                 oauthToken.setAccessToken(result[0].access_token);
-                                oauthToken.setRefreshToken(result[0].refresh_token);
+
                                 oauthToken.setExpiresIn(result[0].expiry_time);
+
+                                oauthToken.setRefreshToken(result[0].refresh_token);
+
+                                oauthToken.setUserMail(result[0].user_mail);
+
                                 resolve(oauthToken);
                             }
+
                             resolve(undefined);
                         });
                     });
@@ -65,24 +94,34 @@ export class DBStore implements TokenStore {
 
     async saveToken(user: UserSignature, token: Token): Promise<void> {
         try {
-            var connection = this.getConnection();
+            var connection = await this.getConnection();
+
             var dbStoreInstance = this;
-            if(token instanceof OAuthToken){
+
+            if (token instanceof OAuthToken) {
                 token.setUserMail(user.getEmail());
-                var sqlQuery = "INSERT INTO oauthtoken(user_mail,client_id,refresh_token,access_token,grant_token,expiry_time) VALUES ('" + user.getEmail() + "'," + "'" + token.getClientID() + "'," + "'" + token.getRefreshToken() + "'," + "'" + token.getAccessToken() + "'," + "'" + token.getGrantToken() + "'," + "'" + token.getExpiresIn() + "')";
-                new Promise<void>(function(resolve, reject){
-                    dbStoreInstance.deleteToken(token).then(function(){
-                        connection.connect(function(err){
-                            if(err){
-                                connection.end();
-                                resolve(undefined);
+                
+                var sqlQuery = "INSERT INTO " + this.tableName + "(id,user_mail,client_id,client_secret,refresh_token,access_token,grant_token,expiry_time,redirect_url) VALUES ?";
+                
+                var values = [
+                    [token.getId(), user.getEmail(), token.getClientId(), token.getClientSecret(), token.getRefreshToken(), token.getAccessToken(), token.getGrantToken(), token.getExpiresIn(), token.getRedirectURL()]
+                ]
+                
+                new Promise<void>(function (resolve, reject) {
+                    dbStoreInstance.deleteToken(token).then(function () {
+                        connection.connect(function (err) {
+                            if (err) {
+                                throw err;
                             }
-                            connection.query(sqlQuery, function(err, result){
+                            connection.query(sqlQuery, [values], function (err, result) {
                                 if (err) {
                                     connection.end();
+                                    
                                     throw err;
                                 }
+                                
                                 connection.end();
+                                
                                 resolve();
                             })
                         })
@@ -96,17 +135,23 @@ export class DBStore implements TokenStore {
 
     async deleteToken(token: Token): Promise<void> {
         try {
-            var connection = this.getConnection();
-            if(token instanceof OAuthToken){
-                var sqlQuery = this.constructDBQuery(token.getUserMail(), token, true);
-                new Promise(function(resolve,reject){
-                    connection.connect(function(err){
-                        if(err) throw err;
-                        connection.query(sqlQuery,function(err,result){
-                            if(err) {
+            var connection = await this.getConnection();
+            
+            if (token instanceof OAuthToken) {
+                
+                var sqlQuery = await this.constructDBQuery(token.getUserMail(), token, true);
+                
+                new Promise(function (resolve, reject) {
+                    connection.connect(function (err) {
+                        if (err) throw err;
+                        
+                        connection.query(sqlQuery, function (err, result) {
+                            if (err) {
                                 throw err;
                             }
+                            
                             connection.end();
+                            
                             resolve(result);
                         })
                     })
@@ -116,56 +161,80 @@ export class DBStore implements TokenStore {
             throw new SDKException(Constants.TOKEN_STORE, Constants.DELETE_TOKEN_DB_ERROR, null, error);
         }
     }
-    async getTokens(): Promise<Token[]> {
+
+    async getTokens(): Promise<Array<Token> | undefined> {
         var tokens: Token[] = [];
+        
         try {
-            var connection = this.getConnection();
-            var sqlQuery = "select * from oauthtoken;";
-            var test = new Promise(function(resolve,reject){
-                connection.connect(function(err){
-                    if(err) throw err;
-                    connection.query(sqlQuery,function(err,result){
-                        if(err) {
+            var connection = await this.getConnection();
+            
+            var sqlQuery = "select * from " + this.tableName + ";";
+            
+            return new Promise(function (resolve, reject) {
+                connection.connect(function (err) {
+                    if (err) throw err;
+                    
+                    connection.query(sqlQuery, function (err, result) {
+                        if (err) {
                             connection.end();
+                            
                             throw err;
                         }
+
                         connection.end();
+                        
                         if (result.length > 0) {
-                            for(let row of result) {
-                                let tokenType = (row.grant_token != null && row.grant_token != Constants.NULL_VALUE && row.grant_token.length > 0) ? TokenType.GRANT : TokenType.REFRESH;
-                                let tokenValue = (tokenType == TokenType.REFRESH) ? row.refresh_token : row.grant_token;
-                                let token = new OAuthToken(row.client_id, "", tokenValue, tokenType);
+                            for (let row of result) {
+                                let grantToken = (row.grant_token !== null && row.grant_token !== Constants.NULL_VALUE && row.grant_token.length > 0) ? row.grant_token : null;
+
+                                let token = new OAuthBuilder().clientId(row.client_id).clientSecret(row.client_secret).refreshToken(row.refresh_token).build();
+
                                 token.setId(row.id);
-                                token.setExpiresIn(row.expiry_time);
+
+                                if (grantToken != null) {
+                                    token.setGrantToken(grantToken);
+                                }
+
                                 token.setUserMail(row.user_mail);
+
                                 token.setAccessToken(row.access_token);
+
+                                token.setExpiresIn(row.expiry_time);
+
+                                token.setRedirectURL(row.redirect_url);
+
                                 tokens.push(token);
                             }
+
                             resolve(tokens);
                         }
-                        resolve(null);
+
+                        resolve(undefined);
                     });
                 });
             });
         } catch (error) {
             throw new SDKException(Constants.TOKEN_STORE, Constants.GET_TOKENS_DB_ERROR, null, error);
         }
-
-        return tokens;
     }
 
-    deleteTokens(): void {
+    async deleteTokens(): Promise<void> {
         try {
-            var connection = this.getConnection();
-            var sqlQuery = "delete from oauthtoken";
-            new Promise(function(resolve,reject){
-                connection.connect(function(err){
-                    if(err) throw err;
-                    connection.query(sqlQuery,function(err,result){
-                        if(err) {
+            var connection = await this.getConnection();
+
+            var sqlQuery = "delete from " + this.tableName + ";";
+            
+            return new Promise(function (resolve, reject) {
+                connection.connect(function (err) {
+                    if (err) throw err;
+                    
+                    connection.query(sqlQuery, function (err, result) {
+                        if (err) {
                             throw err;
                         }
+                        
                         connection.end();
+                        
                         resolve(result);
                     })
                 })
@@ -176,14 +245,21 @@ export class DBStore implements TokenStore {
     }
 
     private constructDBQuery(email: string, token: OAuthToken, isDelete: boolean) {
-        var query = isDelete? "delete from " : "select * from ";
-        query += "oauthtoken " + "where user_mail ='" + email + "' and client_id='" + token.getClientID() + "' and ";
-        if(token.getGrantToken() != null) {
+        if (email == null) {
+            throw new SDKException(Constants.USER_MAIL_NULL_ERROR, Constants.USER_MAIL_NULL_ERROR_MESSAGE);
+        }
+
+        var query = isDelete ? "delete from " : "select * from ";
+
+        query += this.tableName + " where user_mail ='" + email + "' and client_id='" + token.getClientId() + "' and ";
+
+        if (token.getGrantToken() != null) {
             query += "grant_token='" + token.getGrantToken() + "'";
         }
         else {
             query += "refresh_token='" + token.getRefreshToken() + "'";
         }
+
         return query;
     }
 
@@ -199,4 +275,62 @@ export class DBStore implements TokenStore {
         return connection;
     }
 
+    async getTokenById(id: string, token: Token): Promise<Token | undefined> {
+        try {
+            var connection = await this.getConnection();
+
+            if (token instanceof OAuthToken) {
+                var sql = "select * from " + this.tableName + " where id='" + id + "'";
+
+                return new Promise(function (resolve, reject) {
+                    connection.connect(function (err) {
+                        if (err) throw err;
+
+                        connection.query(sql, function (err, result) {
+                            if (err) {
+                                connection.end();
+
+                                throw err;
+                            }
+
+                            connection.end();
+
+                            if (result.length != 0) {
+                                let grantToken = (result[0].grant_token != null && result[0].grant_token !== Constants.NULL_VALUE && result[0].grant_token.length > 0) ? result[0].grant_token : null;
+
+                                token.setClientId(result[0].client_id);
+
+                                token.setClientSecret(result[0].client_secret);
+
+                                token.setRefreshToken(result[0].refresh_token);
+
+                                token.setId(result[0].id);
+
+                                if (grantToken != null) {
+                                    token.setGrantToken(grantToken);
+                                }
+
+                                token.setUserMail(result[0].user_mail);
+
+                                token.setAccessToken(result[0].access_token);
+
+                                token.setExpiresIn(result[0].expiry_time);
+
+                                token.setRedirectURL(result[0].redirect_url);
+
+                                resolve(token);
+                            }
+                            else {
+                                throw new SDKException(Constants.TOKEN_STORE, Constants.GET_TOKEN_BY_ID_DB_ERROR);
+                            }
+
+                            resolve(undefined);
+                        });
+                    });
+                })
+            }
+        } catch (error) {
+            throw new SDKException(Constants.TOKEN_STORE, Constants.GET_TOKEN_DB_ERROR, null, error);
+        }
+    }
 }
